sys_routing, synchronized to lock, refactorings

+++ swiftlets/sys_routing/src/main/java/com/swiftmq/impl/routing/single/connection/stage/Stage.java
@@ -54,7 +54,7 @@ public abstract class Stage {
                 ctx.networkSwiftlet.getConnectionManager().removeConnection(routingConnection.getConnection());
             }
         };
-        ctx.timerSwiftlet.addInstantTimerListener(((Long) ctx.root.getProperty("stage-valid-timeout").getValue()).longValue(), validTimer);
+        ctx.timerSwiftlet.addInstantTimerListener((Long) ctx.root.getProperty("stage-valid-timeout").getValue(), validTimer);
     }
 
     public abstract void process(Request request);
